Add files via upload
If the user is an instructor for a class, he/she could access the file named scores.txt and can add new exam scores to students or add a new student to the class. It can also find the average score of a specific exam. If the user is a student, he/she can access the file named passwords.txt and must enter their username and password. the program checks through the file to see if the username exists and if the password is correct. The student also has the option to change the password, but must enter their username and old password first.